+++ src/Document/Infrastructure/LLMChain/DocumentSearch.php
@@ -13,7 +13,11 @@ use ChronicleKeeper\Settings\Application\SettingsHandler;
 use ChronicleKeeper\Shared\Application\Query\QueryService;
 use ChronicleKeeper\Shared\Infrastructure\LLMChain\EmbeddingCalculator;
 use PhpLlm\LlmChain\Chain\Toolbox\Attribute\AsTool;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
+use function array_filter;
+use function array_map;
+use function array_reduce;
 use function count;
 
 use const PHP_EOL;
@@ -35,6 +39,7 @@ class DocumentSearch
         private readonly SettingsHandler $settingsHandler,
         private readonly QueryService $queryService,
         private readonly RuntimeCollector $runtimeCollector,
+        private readonly UrlGeneratorInterface $router,
     ) {
     }
 
@@ -50,14 +55,14 @@ class DocumentSearch
         $maxResults  = $settings->getChatbotGeneral()->getMaxDocumentResponses();
         $maxDistance = $this->maxDistance ?? $settings->getChatbotTuning()->getDocumentsMaxDistance();
 
-        /** @var list<array{document: Document, content: string, distance: float}> $documents */
-        $documents = $this->queryService->query(new SearchSimilarVectors(
+        /** @var list<array{document: Document, content: string, distance: float}> $vectors */
+        $vectors = $this->queryService->query(new SearchSimilarVectors(
             $this->embeddingCalculator->getSingleEmbedding($search),
             $maxDistance,
             $maxResults,
         ));
 
-        if (count($documents) === 0) {
+        if (count($vectors) === 0) {
             $this->runtimeCollector->addFunctionDebug(
                 new FunctionDebug(
                     'library_documents',
@@ -71,28 +76,60 @@ class DocumentSearch
 
         $debugResponse = [];
 
+        // Group by document
+        /** @var array<string, array{document: Document, content: string, distance: float}> $documents */
+        $documents = array_reduce(
+            $vectors,
+            static function (array $carry, array $vector) {
+                $carry[$vector['document']->getId()] = $vector;
+
+                return $carry;
+            },
+            [],
+        );
+
         $result = 'I have found the following information that are associated to the question:' . PHP_EOL;
         foreach ($documents as $document) {
             $libraryDocument        = $document['document'];
             $documentHierarchyTitle = $libraryDocument->getDirectory()->flattenHierarchyTitle();
 
             $result .= '# Title: ' . $libraryDocument->getTitle() . PHP_EOL;
-            $result .= 'Storage Directory: ' . $documentHierarchyTitle . PHP_EOL;
-            $result .= $document['content'] . PHP_EOL . PHP_EOL;
+            $result .= '- Storage Directory: ' . $documentHierarchyTitle . PHP_EOL;
+            $result .= '- Url: ' . $this->router->generate(
+                'library_document_view',
+                ['document' => $libraryDocument->getId()],
+                UrlGeneratorInterface::ABSOLUTE_URL,
+            ) . PHP_EOL . PHP_EOL;
+
+            $result .= $libraryDocument->getContent() . PHP_EOL . PHP_EOL;
 
             $this->runtimeCollector->addReference(Reference::forDocument($libraryDocument));
 
             $debugResponse[] = [
+                'id' => $libraryDocument->getId(),
                 'document' => $documentHierarchyTitle . '/' . $libraryDocument->getTitle(),
-                'distance' => $document['distance'],
-                'content' => $document['content'],
+                'found_through_vectors' => array_map(
+                    static fn (array $vector) => [
+                        'distance' => $vector['distance'],
+                        'content' => $vector['content'],
+                    ],
+                    array_filter(
+                        $vectors,
+                        static fn (array $vector) => $vector['document']->getId() === $libraryDocument->getId(),
+                    ),
+                ),
             ];
         }
 
         $this->runtimeCollector->addFunctionDebug(
             new FunctionDebug(
                 'library_documents',
-                ['search' => $search, 'maxDistance' => $maxDistance, 'maxResults' => $maxResults],
+                [
+                    'search' => $search,
+                    'maxDistance' => $maxDistance,
+                    'maxResults' => $maxResults,
+                    'vector_chunk_count' => count($vectors),
+                ],
                 $debugResponse,
             ),
         );